services/users/requirements.txt
Click==7.0
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
psycopg2-binary==2.7.7
SQLAlchemy==1.2.15
Werkzeug==0.14.1
Flask-Testing==0.6.2
gunicorn==19.9
coverage==4.5.2
flake8==3.6.0
flask-debugtoolbar==0.10.1
flask-cors==3.0.7
flask-migrate==2.3.1
flask-bcrypt==0.7.1
pyjwt==1.7.1
pytest-flask==0.10.0
pytest-cov==2.6.1

